requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
itsdangerous
bcrypt
aiosqlite
openpyxl
pytesseract
Pillow
